+++ index.html
@@ -12,13 +12,11 @@
     	<header>
             <nav>
                 <div class="nav-wrapper">
-                  <form>
                     <div class="input-field">
                       <input id="searchbar" type="search" required>
                       <label class="label-icon" for="search"><i class="material-icons">search</i></label>
                       <i class="material-icons">close</i>
                     </div>
-                  </form>
                 </div>
             </nav>
             <nav>
